+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 Fixed
 -----
 - Fix ``DictRelatedField`` so it can be used in browsable-API
+- Fix access to subfields of empty ``GroupField`` in Python processes
 
 
 ===================